+++ app/views/dashboard/form/publish/edit.html.erb
@@ -14,8 +14,8 @@
      open_access_version_published_version_rights: permissions_value.call(:published_version_rights, :licence, published_version),
      open_access_version_accepted_version_statement: permissions_value.call(:accepted_version_statement, :publisher_statement, accepted_version),
      open_access_version_published_version_statement: permissions_value.call(:published_version_statement, :publisher_statement, published_version),
-     open_access_version_accepted_version_embargo: permissions_value.call(:accepted_version_embargo, :embargo_end_date, accepted_version),
-     open_access_version_published_version_embargo: permissions_value.call(:published_version_embargo, :embargo_end_date, published_version),
+     open_access_version_accepted_version_embargo: permissions_value.call(:accepted_version_embargo, :embargo_end_date, accepted_version)&.to_s,
+     open_access_version_published_version_embargo: permissions_value.call(:published_version_embargo, :embargo_end_date, published_version)&.to_s,
      open_access_version_current_version: @current_version || accepted_version,
      open_access_version_other_version: @other_version,
      open_access_version_versions_found: params.dig(:oaw_permissions, :versions_found) || @permissions&.versions_found?.to_json,
